+++ lib/browser.d.ts
@@ -246,7 +246,7 @@ export declare const common: {
     endsWithExt: (inode: string, ext: string) => boolean;
     leftPad: (strToPad?: string, outWidth?: number, padChar?: string) => string;
     rightPad: (strToPad?: string, outWidth?: number, padChar?: string) => string;
-    centeredPad: (strToPad?: string, outWidth?: number, padChar?: string) => string;
+    centerPad: (strToPad?: string, outWidth?: number, padChar?: string) => string;
     isNonexistentOrString: (val: any) => val is string;
     isNumberLike: (val: any, allowArrayWith1Num?: boolean) => boolean;
     isBoolean: (val: any) => val is boolean;